First only HTML buildup structure
Hi guys, I'm Jecinta and I'm a website developer as well as designer (Beginners level though but I'm working my way to becoming a pro). The above code is HTML and this is my first trial on website creation. I only started recently and would daily update my repository with new codes. So, I tried creating my first website using only HTML, and it was fun, didn't expect it to be this amazing though. Stay tuned for more updates are yet to come.